+++ tests/DeprecatedCsrfMiddlewareTest.php
@@ -6,18 +6,17 @@ namespace Yiisoft\Csrf\Tests;
 
 use Nyholm\Psr7\Factory\Psr17Factory;
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Csrf\CsrfTokenMiddleware;
+use Yiisoft\Csrf\CsrfMiddleware;
 use Yiisoft\Csrf\Synchronizer\Generator\RandomCsrfTokenGenerator;
 use Yiisoft\Csrf\Synchronizer\SynchronizerCsrfToken;
 use Yiisoft\Csrf\Tests\Synchronizer\Storage\MockCsrfTokenStorage;
-use Yiisoft\Http\Method;
 
-final class CsrfMiddlewareTest extends TestCase
+final class DeprecatedCsrfMiddlewareTest extends TestCase
 {
     public function testDefaultParameterName(): void
     {
         $middleware = $this->createMiddleware();
-        $this->assertSame(CsrfTokenMiddleware::PARAMETER_NAME, $middleware->getParameterName());
+        $this->assertSame(CsrfMiddleware::PARAMETER_NAME, $middleware->getParameterName());
     }
 
     public function testGetParameterName(): void
@@ -31,7 +30,7 @@ final class CsrfMiddlewareTest extends TestCase
     public function testDefaultHeaderName(): void
     {
         $middleware = $this->createMiddleware();
-        $this->assertSame(CsrfTokenMiddleware::HEADER_NAME, $middleware->getHeaderName());
+        $this->assertSame(CsrfMiddleware::HEADER_NAME, $middleware->getHeaderName());
     }
 
     public function testGetHeaderName(): void
@@ -47,12 +46,11 @@ final class CsrfMiddlewareTest extends TestCase
         $original = $this->createMiddleware();
         $this->assertNotSame($original, $original->withHeaderName('csrf'));
         $this->assertNotSame($original, $original->withParameterName('csrf'));
-        $this->assertNotSame($original, $original->withSafeMethods([Method::HEAD]));
     }
 
-    private function createMiddleware(): CsrfTokenMiddleware
+    private function createMiddleware(): CsrfMiddleware
     {
-        return new CsrfTokenMiddleware(
+        return new CsrfMiddleware(
             new Psr17Factory(),
             new SynchronizerCsrfToken(
                 new RandomCsrfTokenGenerator(),