+++ tests/DeprecatedTokenCsrfMiddlewareTest.php
@@ -11,39 +11,18 @@ use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Yiisoft\Csrf\CsrfTokenMiddleware;
+use Yiisoft\Csrf\CsrfMiddleware;
 use Yiisoft\Csrf\CsrfTokenInterface;
 use Yiisoft\Csrf\MaskedCsrfToken;
 use Yiisoft\Http\Method;
 use Yiisoft\Http\Status;
 use Yiisoft\Security\Random;
 
-abstract class TokenCsrfMiddlewareTest extends TestCase
+abstract class DeprecatedTokenCsrfMiddlewareTest extends TestCase
 {
     private const PARAM_NAME = 'csrf';
 
     private string $token;
-
-    public function testGetIsAlwaysAllowed(): void
-    {
-        $middleware = $this->createCsrfMiddleware();
-        $response = $middleware->process($this->createServerRequest(Method::GET), $this->createRequestHandler());
-        $this->assertEquals(200, $response->getStatusCode());
-    }
-
-    public function testHeadIsAlwaysAllowed(): void
-    {
-        $middleware = $this->createCsrfMiddleware();
-        $response = $middleware->process($this->createServerRequest(Method::HEAD), $this->createRequestHandler());
-        $this->assertEquals(200, $response->getStatusCode());
-    }
-
-    public function testOptionsIsAlwaysAllowed(): void
-    {
-        $middleware = $this->createCsrfMiddleware();
-        $response = $middleware->process($this->createServerRequest(Method::OPTIONS), $this->createRequestHandler());
-        $this->assertEquals(200, $response->getStatusCode());
-    }
 
     public function testValidTokenInBodyPostRequestResultIn200(): void
     {
@@ -100,6 +79,13 @@ abstract class TokenCsrfMiddlewareTest extends TestCase
         $this->assertEquals(200, $response->getStatusCode());
     }
 
+    public function testGetIsAlwaysAllowed(): void
+    {
+        $middleware = $this->createCsrfMiddleware();
+        $response = $middleware->process($this->createServerRequest(Method::GET), $this->createRequestHandler());
+        $this->assertEquals(200, $response->getStatusCode());
+    }
+
     public function testInvalidTokenResultIn422(): void
     {
         $middleware = $this->createCsrfMiddleware();
@@ -144,42 +130,6 @@ abstract class TokenCsrfMiddlewareTest extends TestCase
         $this->assertEquals(Status::UNPROCESSABLE_ENTITY, $response->getStatusCode());
     }
 
-    public function testUnsafeMethodPostRequestResultIn422(): void
-    {
-        $middleware = $this->createCsrfMiddleware();
-        $response = $middleware->process(
-            $this->createServerRequest(Method::POST),
-            $this->createRequestHandler()
-        );
-        $this->assertEquals(Status::TEXTS[Status::UNPROCESSABLE_ENTITY], $response->getBody());
-        $this->assertEquals(Status::UNPROCESSABLE_ENTITY, $response->getStatusCode());
-    }
-
-    public function testCustomSafeOptionsRequestResultIn200(): void
-    {
-        $middleware = $this
-            ->createCsrfMiddleware()
-            ->withSafeMethods([Method::OPTIONS]);
-        $response = $middleware->process(
-            $this->createServerRequest(Method::OPTIONS),
-            $this->createRequestHandler()
-        );
-        $this->assertEquals(200, $response->getStatusCode());
-    }
-
-    public function testCustomUnsafeMethodGetRequestResultIn422(): void
-    {
-        $middleware = $this
-            ->createCsrfMiddleware()
-            ->withSafeMethods([Method::OPTIONS]);
-        $response = $middleware->process(
-            $this->createServerRequest(Method::GET),
-            $this->createRequestHandler()
-        );
-        $this->assertEquals(Status::TEXTS[Status::UNPROCESSABLE_ENTITY], $response->getBody());
-        $this->assertEquals(Status::UNPROCESSABLE_ENTITY, $response->getStatusCode());
-    }
-
     private function createServerRequest(
         string $method = Method::POST,
         array $bodyParams = [],
@@ -206,7 +156,7 @@ abstract class TokenCsrfMiddlewareTest extends TestCase
 
     private function createPostServerRequestWithHeaderToken(
         string $token,
-        string $headerName = CsrfTokenMiddleware::HEADER_NAME
+        string $headerName = CsrfMiddleware::HEADER_NAME
     ): ServerRequestInterface {
         return $this->createServerRequest(Method::POST, [], [
             $headerName => $token,
@@ -233,11 +183,11 @@ abstract class TokenCsrfMiddlewareTest extends TestCase
     protected function createCsrfMiddleware(
         ?CsrfTokenInterface $csrfToken = null,
         RequestHandlerInterface $failureHandler = null
-    ): CsrfTokenMiddleware {
+    ): CsrfMiddleware {
         $csrfToken = new MaskedCsrfToken($csrfToken ?? $this->createCsrfToken());
         $this->token = $csrfToken->getValue();
 
-        $middleware = new CsrfTokenMiddleware(new Psr17Factory(), $csrfToken, $failureHandler);
+        $middleware = new CsrfMiddleware(new Psr17Factory(), $csrfToken, $failureHandler);
 
         return $middleware->withParameterName(self::PARAM_NAME);
     }